Fixed error when constance needs a hidden field
Recently django-constance has added a way to prevent changes made by someone else at the same time, this requires to render a hidden field to validate it, you can see the change here:

https://github.com/jezdez/django-constance/blob/abd42df3046740e77a443acf479bdfb4e529b251/constance/templates/admin/constance/change_list.html#L40

I have added the same code, using grappelli style to copy the same behavior.

It was very hard to find this was causing the problem, so I would recommend to find a way to remove this template override, it is very hard to find any issue, and also should be hard to keep up with the changes on the original library.

From what I saw this was created just to support the calendar widget, maybe there is a different approach to keep the same support without having to maintain a copy of the template here.

It would be very nice if you make a quick pypi release with this fix.

Thanks